data/manager/requirements.txt
paho-mqtt
docker
paramiko